Nicer errors when materialized path does not exist

+++ lib/materialize.nix
@@ -1,9 +1,21 @@
 { pkgs, nix, runCommand, checkMaterialization }@defaults:
 { sha256
-, sha256Arg
-, materialized
-, reasonNotSafe
+, sha256Arg      # Name of the sha256 argument for more meaningful
+                 # error messages when checking the materialization.
+, materialized   # null or path where to find materialized version of
+                 # the output. If this is set but does not exist
+                 # the derivation will fail but with a message
+                 # advising how to populate it.
+, reasonNotSafe  # Some times there a reasont the derivation will
+                 # not produce output that can be safely materialized.
+                 # Set this to a string explaining why and materialization
+                 # will not be used (if sha256 was set an error will be
+                 # displayed including the reasonNotSafe string).
 , checkMaterialization ? defaults.checkMaterialization
+                 # When checkMaterialization is set the derivation
+                 # will be calculated the slow way (without using `sha256`
+                 # and `materialized`) the result will be used to check
+                 # `sha256` and `materialized` (if set).
 }: derivation:
 
 let
@@ -57,19 +69,28 @@ let
           echo ${sha256Arg} used for ${name} is correct
         fi
       '')
-    + (pkgs.lib.optionalString (materialized != null) ''
-        if diff -qr ${materialized} ${calculateNoHash} &>/dev/null; then
-          echo materialized nix used for ${name} is correct
+    + (
+      if materialized != null && !__pathExists materialized
+        then ''
+          echo materialized nix used for ${name} is missing. To fix run :
+          echo cp -r ${calculateNoHash} ${toString materialized}
+          echo chmod -R +w ${toString materialized}
+          false
+        ''
         else
-          echo Changes to plan not reflected in materialized nix for ${name}
-          diff -ru ${materialized} ${calculateNoHash}
-        fi
-      '')
-    + ''
-        cp -r ${unchecked} $out
-        # Make sure output files can be removed from the sandbox
-        chmod -R +w $out
-      ''
+          (pkgs.lib.optionalString (materialized != null && __pathExists materialized) ''
+            if diff -qr ${materialized} ${calculateNoHash} &>/dev/null; then
+              echo materialized nix used for ${name} is correct
+            else
+              echo Changes to plan not reflected in materialized nix for ${name}
+              diff -ru ${materialized} ${calculateNoHash}
+            fi
+          '')
+        + ''
+            cp -r ${unchecked} $out
+            # Make sure output files can be removed from the sandbox
+            chmod -R +w $out
+          '')
   );
 
   hashArgs = {
@@ -81,15 +102,19 @@ let
   calculateUseHash = derivation.overrideAttrs (_: hashArgs);
   calculateUseAll = 
     # Skip right to expectedPath if it already exists
-    if materialized != null && builtins.pathExists materialized
-      then runCommand name hashArgs ''
-        cp -r ${materialized} $out
-        # Make sure output files can be removed from the sandbox
-        chmod -R +w $out
-      ''
+    if materialized != null
+      then
+        assert __pathExists materialized;
+        runCommand name hashArgs ''
+          cp -r ${materialized} $out
+          # Make sure output files can be removed from the sandbox
+          chmod -R +w $out
+        ''
       else calculateUseHash;
 
 in
-  if checkMaterialization
+  # Use the checked version if requested or if the `materialized` version
+  # is missing (perhaps deleted or not created yet).
+  if checkMaterialization || (materialized != null && !__pathExists materialized)
     then checked
     else unchecked
